Fixed a scaling problem that I now recognize as a disagreement between signal power and equivalent DC voltage (a.k.a RMS).
Since the DBX process is specified as a power-based expansion
(power dB in/power dB out), the envelope follower should return
power estimates and not RMS estimates. The difference between the
two estimates comes down to removing the sqrt() call inside the
estimator loop.

Not only is this easier to compare with the mathematical description
of the processor, it's also slightly less computationally intensive!

A side effect of this change is that the internal "ratio" variable
needs to be halved, since it will be used to drive a scalar gain
value, and hence, must become a voltage-domain scalar.

The default thresholds might still need to be adjusted.